refactor: extract topics toggle into component module
