+++ foundation_cms/static/js/pages/expert_directory_page.js
@@ -1,16 +1,3 @@
-document.addEventListener("click", (e) => {
-  const btn = e.target.closest(".expert-card__topics-toggle");
-  if (!btn) return;
+import { initTopicsToggle } from "../components/expert_directory_page/topics_toggle";
 
-  const topicsContainer = btn.closest(".expert-card__topics");
-  if (!topicsContainer) return;
-
-  const expanded = topicsContainer.classList.toggle(
-    "expert-card__topics--expanded",
-  );
-  btn.setAttribute("aria-expanded", String(expanded));
-  btn.setAttribute(
-    "aria-label",
-    expanded ? btn.dataset.labelCollapse : btn.dataset.labelExpand,
-  );
-});
+initTopicsToggle();
